+++ Utility.h
@@ -2,7 +2,7 @@
 #define UTILITY_H
 
 /** \file
-  * File containing useful functions.
+  * Header file of utility.
   */
 
 #include <tsk/libtsk.h>
@@ -12,6 +12,6 @@ uint16_t read16Bit(TSK_ENDIAN_ENUM endian, uint8_t *arr);
 
 uint32_t read32Bit(TSK_ENDIAN_ENUM endian, uint8_t *arr);
 
-uint32_t read64Bit(TSK_ENDIAN_ENUM endian, uint8_t *arr);
+uint64_t read64Bit(TSK_ENDIAN_ENUM endian, uint8_t *arr);
 
 #endif
